pyproject.toml
[project]
name = "soulpage-assignment-conversational-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ddgs>=9.10.0",
    "duckduckgo-search>=8.1.1",
    "langchain-community>=0.4.1",
    "langchain-openai~=1.1.0",
    "langchain~=1.2.0",
    "langgraph>=1.0.7",
    "python-dotenv~=1.2.0",
    "wikipedia~=1.4.0",
]
